:lipstick: Move script tests into a subpackage
